requirements.txt
Django>=4.2.0
Pillow>=11.0.0
opencv-python>=4.8.0
numpy>=1.24.0
